only creator can update or delete playlist

+++ libs/frontend/features/src/lib/playlist/playlist-detail/playlist-detail.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { IPlaylist, ISong } from '@avans-nx-songlibrary/api';
+import { IPlaylist, ISong, IUser } from '@avans-nx-songlibrary/api';
 import { Subscription } from 'rxjs';
 import { ActivatedRoute, Router } from '@angular/router';
 import { PlaylistService } from '../playlist.service';
 import { Location } from '@angular/common';
+import { LoginService } from '../../login/login.service';
 
 @Component({
     selector: 'playlist-detail',
@@ -12,12 +13,18 @@ import { Location } from '@angular/common';
 export class PlaylistDetailComponent implements OnInit, OnDestroy {
     playlist: IPlaylist | undefined;
     songs: ISong[] = [];
+    user: IUser | undefined
 
     id: string | null = null;
     private sub: Subscription | null = null;
-    constructor(private playlistService: PlaylistService, private route: ActivatedRoute, private router: Router, private location: Location) {}
+    constructor(private playlistService: PlaylistService, private route: ActivatedRoute, private router: Router, private location: Location, private loginService: LoginService) {}
 
     ngOnInit(): void {
+        this.loginService.currentUser.subscribe((user) => {
+            if (user) {
+                this.user = user;
+            }
+        });
         this.route.paramMap.subscribe((params) => {
             this.id = params.get('id');
             if (this.id) {
@@ -40,20 +47,26 @@ export class PlaylistDetailComponent implements OnInit, OnDestroy {
     }
 
     deletePlaylist(playlist: IPlaylist): void {
-        this.playlistService.delete(playlist).subscribe({
-            next: () => {
-                this.router.navigate(['playlist-list']);
-            },
-            error: (err) => console.error('Failed to delete playlist', err)
-        });
+        if(playlist.creator == this.user) {
+            this.playlistService.delete(playlist).subscribe({
+                next: () => {
+                    this.router.navigate(['playlist-list']);
+                },
+                error: (err) => console.error('Failed to delete playlist', err)
+            });
+        }
+        
     }
     
     updatePlaylist(): void {
-        if (this.playlist) {
+        if(this.playlist && this.playlist.creator == this.user) {
+            if (this.playlist) {
             this.playlistService.update(this.playlist).subscribe((updatedPlaylist) => {
                 this.playlist = updatedPlaylist;
             });
         }
+        }
+        
     }
 
     goBack(): void {
